refactored code to use promses
